Added a util class for useful functions
