+++ app/src/test/java/ch/sdp/vibester/UtilTests.kt
@@ -1,0 +1,18 @@
+package ch.sdp.vibester
+
+import ch.sdp.vibester.util.Util
+import org.junit.Test
+
+import org.junit.Assert.*
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class UtilTests {
+    @Test
+    fun newIdIsCorrectLength() {
+        assertEquals(Util.createNewId().length, 10)
+    }
+}
